% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baysCO16ported.R
\name{besseliln}
\alias{besseliln}
\title{my_mode}
\usage{
besseliln(nu, z)
}
\arguments{
\item{nu}{numeric; The order (maybe fractional!) of the corresponding Bessel function.}

\item{z}{numeric, ≥ 0.}
}
\description{
log of base::besselI
}
\author{
Wanja Mössing
}
