% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.blinktreatment.R
\name{blink.detect}
\alias{blink.detect}
\title{Detect blinks in (Eyelink 1000) pupil dilation data}
\usage{
blink.detect(
  pddt,
  minDilation = 500,
  maxDeltaDilation = 5,
  TrialCol = "Trial",
  SR_Blink_Col = NA,
  use_SR = FALSE,
  verbose = TRUE,
  expandblinks = 0
)
}
\arguments{
\item{pddt}{a pupil dilation \code{data.table} of a single participant
containing at least the following columns:
\itemize{
\item{Dil} containing the dilation samples
\item{Time} time stamp in ms associated with each sample
\item{Trial} indicating to which trial the current sample belongs
}
An optional column coding Blinks in 0 or 1 (as provided by the edfR package)
can be used as basis for blink detection and/or comparison of the two
methods.}

\item{minDilation}{Dilations lower than this are considered a 'Blink'}

\item{maxDeltaDilation}{By how much are samples allowed to differ from the
following sample? Any sample exceeding this threshold is considered a blink.}

\item{TrialCol}{\code{char}, the name of the column coding trials (default:
\code{'Trial'})}

\item{SR_Blink_Col}{\code{char}, the name of the column containing the
online-detected blinks. Jason Hubbard's \code{edfR} provides a column
'Blink', which is 1 for blink samples and 0 for non-blink samples.}

\item{use_SR}{\code{logical}, ignore minDilation and maxDeltaDilation and
base blink detection solely on SR-Research's online detection (default:
\code{FALSE})}

\item{verbose}{\code{logical}, if \code{TRUE} (default), will print some
statistics comparing on- and offline detection, if SR_Blink_Col is
available.}

\item{expandblinks}{\code{integer}, just before and after a blink, the pupil
dilation is often affected. expand blinks by N ms. Default is 0.}
}
\value{
Returns the same data.table with column \code{Dil} set to \code{NA}
for blink samples.
}
\description{
Detect blinks in (Eyelink 1000) pupil dilation data
}
\author{
Wanja Mössing
}
