% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.blinktreatment.R
\name{blink.interolate}
\alias{blink.interolate}
\alias{blink.interpolate}
\title{Interpolate blinks in Eyetracking data measured with Eyelink 1000+}
\usage{
blink.interpolate(pddt, type = "spline")
}
\arguments{
\item{pddt}{A pupil dilation data.table, containing at least the columns:
\code{Dil} and \code{Trial} for linear interpolation or \code{Dil} for
spline interpolation.}

\item{type}{character. Type of interpolation. Can be 'spline' (default) or
'linear'}
}
\value{
Returns the same data.table with column \code{Dil} interpolated for
blink periods and an additional column \code{IthBlink} counting blinks per
subject (assuming you're running this on single subjects).
}
\description{
interpolates blinks in pupil dilation data, as detected by
\code{blink.detect}. Currently supports linear and spline interpolation.
Linear interpolation is based on \code{pR::interpolateblinks} by Hedderik
van Rijn (see \code{github.com/hedderik/pR}). The cubic spline interpolation
implements the interpolation part of Mathot, 2013. Note that this function
does not run blink detection as described in that paper.
}
\author{
Wanja Mössing
}
