% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baysCO16ported.R
\name{CO16_fit}
\alias{CO16_fit}
\title{CO16_FIT}
\usage{
CO16_fit(X, TT = NULL, NT = NULL)
}
\arguments{
\item{X}{= [n*1,1] vector of responses}

\item{TT}{= [n*1,1] column vector of Target orientations}

\item{NT}{= [n*1, m] matrix of non-target values}
}
\description{
This is a ported version of Paul Bays' Matlab code published at http://www.paulbays.com/code/CO16/
  ----------------------------------
  Original documentation
  ----------------------------------

  This code is released under a GNU General Public License:
  feel free to use and adapt these functions as you like, but credit should be given to Paul Bays if they contribute to your work, by citing:
  Schneegans S & Bays PM. No fixed item limit in visuospatial working memory. Cortex 83: 181-193 (2016)

  Returns maximum likelihood parameters B (res[1]) for a mixture model describing
  recall responses X in terms of target TT, non-target NT, and uniform
  responses. Inputs should be in radians, -\code{pi} <= X < \code{pi}. Fitting is based
  on an EM algorithm with multiple starting parameters.

  \code{CO16_fit(X, TT, NT)} returns a three-element list with:
  1. vector \code{B c(K, pT, pN, pU)}, where K is
    the concentration parameter of a Von Mises distribution capturing
    response variability, pT is the probability of responding with the
    target value, pN the probability of responding with a non-target
    value, and pU the probability of responding "randomly".

  2. vector \code{LL} additionally returns the log likelihood LL.

  3. vector \code{W} additionally returns a weight matrix of
    trial-by-trial posterior probabilities that responses come from each of
    the three mixture components. Each row of W corresponds to a separate
    trial and is of the form \code{c(wT, wN, wU)}, corresponding to the probability
    the response comes from the target, non-target or uniform response
    distributions, respectively.
}
\references{
Schneegans S & Bays PM. No fixed item limit in visuospatial working
  memory. Cortex 83: 181-193 (2016),
  Bays PM, Catalao RFG & Husain M. The precision of visual working
  memory is set by allocation of a shared resource. Journal of Vision
  9(10): 7, 1-11 (2009)
}
\author{
Paul Bays, R port by Wanja Moessing
}
