% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baysCO16ported.R
\name{CO16_function}
\alias{CO16_function}
\title{CO16_function}
\usage{
CO16_function(X, TT = NULL, NT = NULL, B_start = NULL)
}
\arguments{
\item{X}{= [n*1,1] vector of responses}

\item{TT}{= [n*1,1] column vector of Target orientations}

\item{NT}{= [n*1, m] matrix of non-target values}

\item{B_start}{starting values}
}
\description{
sub-function of CO16_fit
}
\author{
Paul Bays, R port by Wanja Moessing
}
