% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coord2PsyPy.R
\name{Coord2PsyPy}
\alias{Coord2PsyPy}
\title{Coord2PsyPy}
\usage{
Coord2PsyPy(x, y, xmax = 1024, ymax = 768)
}
\arguments{
\item{x}{positive integer x-coordinate}

\item{y}{positive integer y-coordinate}

\item{xmax}{maximum x pixel coordinate in original picture}

\item{ymax}{maximum y pixel coordinate in original picture}
}
\value{
a vector c(x,y) with psychopy coordinates
}
\description{
\code{Coord2PsyPy} returns c(x,y) in psychopy format
Many devices assume that the top-left corner of an image (or screen) has the
coordinates \code{c(x=0, y=0)} and \code{c(x=max, y=max)} in the bottom right
corner. Psychopy assumes that \code{c(x=0, y=0)} is in the centre. This
function takes 'standard' coordinates and transforms them to Psychopy
coordinates.
}
\examples{
pp.pix = Coord2PsyPy(100, 200, 1024, 768)

}
\author{
Wanja Mössing
}
