% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.visang.R
\name{deg2pix}
\alias{deg2pix}
\title{deg2pix}
\usage{
deg2pix(n_deg, screen_width = 52.2, distance = 86, res_x = 1920)
}
\arguments{
\item{n_deg}{how many degree?}

\item{screen_width}{width of the screen in cm}

\item{distance}{viewing distance in cm}

\item{res_x}{horizontal screen resolution in pixels}
}
\value{
n_deg in pixels
}
\description{
converts degree to pixels, defaults are set to the parameters in the buschlab.
}
\examples{
deg2pix(1)

}
\seealso{
pix2deg wmR.angdiff
}
\author{
Wanja Mössing
}
