% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baysCO16ported.R
\name{DT2CO16}
\alias{DT2CO16}
\title{DT2C016}
\usage{
DT2CO16(DT, dt.X = "^X", dt.TT = "^TT", dt.NT = "^NT",
  dt.Trial = "^Trial", type = "rad")
}
\arguments{
\item{DT}{a \code{data.table} containg one column with X (responses) data,
none or one with TT (target) data, and none or multiple with NT (non-target)
data. Data should be in radians (default) or degree (see \code{type})}

\item{dt.X}{pattern that will match the column containing responses. Default
matches columns starting with 'X'}

\item{dt.TT}{pattern that will match the column containing targets. Default
matches columns starting with 'TT'}

\item{dt.NT}{pattern that will match the column containing distractors. Default
matches columns starting with 'NT'}

\item{dt.Trial}{pattern that will match the column containing Trials. Default
matches columns starting with 'Trial'. Ignored if missing.}

\item{type}{Are the input data in degree ('deg') or in radian ('rad',
default). If 'deg' is specified, data will be converted to radian.}
}
\description{
Currently, the ported CO16 functions require matrices instead of
\code{data.tables} (or \code{data.frames}). This function takes care of
extracting the right columns and converts them to matrices. This function is
called internally in \code{CO16_fit}.
}
\author{
Wanja Mössing
}
