% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.EEG.R
\name{eeglabCSV2eegUtils}
\alias{eeglabCSV2eegUtils}
\title{Importing eeg tables from EEGlab}
\usage{
eeglabCSV2eegUtils(DT, channel_locfile = NULL)
}
\arguments{
\item{DT}{a \code{data.table} with exactly the columns specified for the csv above}

\item{channel_locfile}{a .txt file with spherical coordinates for your channels.
defaults to the scalp-channels of the custom Easycap montage used in our lab
at the WWU Münster, Germany.}
}
\value{
an \code{eegUtils} object
}
\description{
This is the only way to import eeglab data to eegUtils that
really worked for me. This assumes that you wrangled your data in eeglab/matlab
into a long table and then exported it as a csv. The csv should have the
following format:
\itemize{
\item one columns per Electrode, column names should be electrode names.
\item one column coding participant ID, column name must be 'participant_id' (can also be used for trials)
\item one column coding time, called 'time'
\item one column with a unique value representing sampling rate, called 'srate'
\item one column with a unique value representing the used reference, called 'reference'
}
How to get your data into that format likely depends heavily on the specific
 situation. Thus, there's no matlab-pendant to this function. Nevertheless,
in the examples section below, you can find one possible way of accomplishing
 this in Matlab.

Note that this function can currently read neither the EEG.events structure
(good luck flattening that to a csv...), nor the EEG.chanlocs structure.
Chanlocs are substituted by reading the channel locations from a .txt file
instead.


Note: depends on Matt Craddock's eegUtils package, which is not available on
CRAN. Use \code{remotes::install_github("craddm/eegUtils")} instead.
}
\examples{
# Matlab first, R below
\dotrun{
#\%\%\%\%\% EXPORTING IN MATLAB \%\%\%\%\% (usage in R below)
# EEG.bsldata are baselined data (64x1900x30; chan x time x id)
# stack the third dimension, so this becomes 2D
#FLAT = reshape(EEG.bsldata, size(EEG.bsldata,1),[],1);
#
#\% make this a table
#FLAT = array2table(FLAT');
#
#\% set column names to channel names
#CHANS = {EEG.chanlocs.labels};
#FLAT.Properties.VariableNames = CHANS;
#
#\% Add column that codes ID
#ID = repelem(1:size(EEG.bsldata, 3), size(EEG.bsldata, 2))';
#                   FLAT.participant_id = ID;
#
#\% add column that codes time
#TIME = repmat(EEG.times, 1, size(EEG.bsldata,3))';
#FLAT.time = TIME;
#
#\% add column with unique sampling rate value
#FLAT.srate = repelem(EEG.srate, height(FLAT), 1);
#
#\% add column with unique reference name
#FLAT.reference = repelem('average', height(FLAT), 1);
#
#\% write table to csv
#writetable(FLAT, 'foo.csv')
#
#-----------------------------
#-----------------------------
#-----------------------------
###### Importing in R #######
wmR::libraries(data.table, eegUtils, ggplot2, wmR)
EEG <- fread('foo.csv')
EEG <- EEG = eeglabCSV2eegUtils(EEG)
subEEG = select_times(EEG, c(0,100))
ggplot(subEEG  , aes(x = x, y = y, fill = amplitude)) + geom_topo()
}
}
\seealso{
eegUtils txt_elec_2_eegUtils
}
\author{
Wanja Mössing
}
