% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.fair_distribution.R
\name{fair_distribution}
\alias{fair_distribution}
\title{Generates a somewhat fair distribution of topics for a given set of groups with ordered preferences per group.}
\usage{
fair_distribution(DT)
}
\arguments{
\item{DT}{a \code{data.table} where each column represents one group and the rows contain their choices in the preferred order.}
}
\value{
Returns a 1-row \code{data.table} with the final topic per group.
}
\description{
Generates a somewhat fair distribution of topics for a given set of groups with ordered preferences per group.
}
\examples{
require(data.table)
DT <- data.table(group1 = sample(5), group2 = sample(5), group3 = sample(5),
                 group4 = sample(5), group5 = sample(5))
fair_distribution(DT)

}
\author{
Wanja Mössing
}
