% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.oneliner.R
\name{libraries}
\alias{libraries}
\title{Load multiple libraries with one command}
\usage{
libraries(...)
}
\arguments{
\item{...}{comma-separated library names without or with quotes. Or a
character vector of library names.}
}
\value{
Nothing. Unloadable packs will throw errors.
}
\description{
loads multiple libraries in one line without massive printing
to console.
}
\examples{
wmR::libraries(data.table, lme4, ggplot2)
wmR::libraries('data.table', 'lme4', 'ggplot2')
packs <- c('data.table', 'lme4', 'ggplot2')
wmR::libraries(packs)

}
\seealso{
\code{library()}
}
\author{
Wanja Mössing
}
