% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.oneliner.R
\name{NaNs}
\alias{NaNs}
\title{NaNs}
\usage{
NaNs(n)
}
\arguments{
\item{n}{an \code{integer}. Length of vector.}
}
\value{
Returns a 1-row \code{vector} with N NaNs
}
\description{
Generates a vector of NaNs with variable length
}
\examples{
mynans <- NaNs(100)

library(data.table)
DT <- data.table(A = NaNs(50))

}
\seealso{
\code{nastr(), NAs()}
}
\author{
Wanja Mössing
}
