% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.oneliner.R
\name{NAs}
\alias{NAs}
\title{NAs}
\usage{
NAs(n)
}
\arguments{
\item{n}{an \code{integer}. Length of vector.}
}
\value{
Returns a 1-row \code{vector} with N NaNs
}
\description{
Generates a vector of NAs with variable length
}
\examples{
mynas <- NAs(100)

library(data.table)
DT <- data.table(A = NAs(50))

}
\seealso{
\code{NaNs(), nastr()}
}
\author{
Wanja Mössing
}
