% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.oneliner.R
\name{nastr}
\alias{nastr}
\title{nastr}
\usage{
nastr(n)
}
\arguments{
\item{n}{an \code{integer}. Length of vector.}
}
\value{
Returns a 1-row \code{vector} with N NaNs
}
\description{
Generates a vector of NA_character_ with variable length.
Useful for preallocating a character variable.
}
\examples{
mynas <- nastr(100)

library(data.table)
DT <- data.table(A = nastr(50))

}
\seealso{
\code{NaNs(), NAs()}
}
\author{
Wanja Mössing
}
