% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.oneliner.R
\name{pcswitch}
\alias{pcswitch}
\title{pcswitch}
\usage{
pcswitch(Names = nastr(3), ..., append = "")
}
\arguments{
\item{Names}{a vector of Names to be compared to the \code{computername()}}

\item{...}{number of elements to be evaluated upon a match.
Order must match Names. That is, if \code{Names[1] == computername()}, the
first element in \code{...} will be evaluated}

\item{append}{append this string to the result. silently ignored, if the
result isn't a string}
}
\value{
Returns the result of whatever it is supposed to evaluate
}
\description{
Does something specific to the host it's running on.
}
\examples{
rootdir <- pcswitch(c('LinuxServer', 'WinLaptop', 'WinStation'),
                      append = 'User',
                      '/', 'C:/', 'F:/')

rootdir <- pcswitch(c('LinuxServer', 'WinLaptop', 'WinStation'),
                      mean(1:99), 77, 'F:/')


}
\author{
Wanja Mössing
}
