% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.visang.R
\name{pix2deg}
\alias{pix2deg}
\title{pix2deg}
\usage{
pix2deg(n_pix, screen_width = 52.2, distance = 86, res_x = 1920)
}
\arguments{
\item{n_pix}{how many pixels?}

\item{screen_width}{width of the screen in cm}

\item{distance}{viewing distance in cm}

\item{res_x}{horizontal screen resolution in pixels}
}
\value{
n_pix in degree visual angle
}
\description{
converts pixels to degree, defaults are set to the parameters in the buschlab.
}
\examples{
pix2deg(1)

}
\seealso{
deg2pix wmR.angdiff
}
\author{
Wanja Mössing
}
