% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.plotSEM.R
\name{sem}
\alias{sem}
\alias{plotSEMminmax}
\alias{plotSEM}
\title{Calculate default between group standard error of the mean.}
\usage{
sem(x, na.rm = FALSE)

plotSEMminmax(x, na.rm = FALSE)

plotSEM(x, na.rm = FALSE)
}
\arguments{
\item{x}{a \code{numeric vector}}

\item{na.rm}{remove NaNs? Default is FALSE.}
}
\value{
a numeric vector with values for errorbars
}
\description{
\code{sem} returns the standard error of the mean
\deqn{\sigma_\bar{x} = \frac{\sigma}{\sqrt{n}}}
\code{plotSEM} is useful for ggplot. It returns a vector with:
mean-SEM, mean, mean+SEM; so very handy for errorbars.
\code{plotSEMminmax} additionally returns minimum and maximum for more elaborate errorbars.
For within group comparisons, use Cosineau-Morey CIs/SEs (e.g., \code{Rmisc::summarySEwithin}).
Use \code{stat_CosineauMoreyCI} as a layer for ggplot.
}
\examples{
x <- rnorm(1000, mean = 5, sd = 90)

sem(x)
plotSEM(x)
plotSEMminmax(x)

# example usage in ggplot
require(ggplot2)
dat <- data.frame(measure = x, condition = c('A', 'B'))

ggplot(data = dat, aes(x = condition, y = measure)) +
       stat_summary(fun.data = plotSEMminmax,
                    geom = "boxplot") +
       ggtitle("Min, Mean-1SEM, Mean, Mean+1SEM, Max")

ggplot(data = dat, aes(x = condition, y = measure)) +
       stat_summary(fun.data = plotSEM,
                    geom = "pointrange") +
       ggtitle("Min, Mean-1SEM, Mean, Mean+1SEM, Max")

}
\seealso{
Rmisc::summarySEwithin, plotSEM, stat_CosineauMoreyCI, plotSEMminmax
}
\author{
Wanja Mössing
}
