% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.plotSEM.R
\docType{data}
\name{stat_CosineauMoreyCI}
\alias{stat_CosineauMoreyCI}
\alias{stat_CosineauMoreySE}
\alias{.CosiMoreCIs}
\alias{.CosiMoreSEs}
\title{Add layer with Cosineau-Morey 95\% within group CIs/ SE to ggplot}
\format{An object of class \code{CosiMoreCIs} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 6.}
\usage{
stat_CosineauMoreyCI(mapping = NULL, data = NULL,
  geom = "pointrange", position = "identity", show.legend = NA,
  inherit.aes = TRUE, na.rm = TRUE, conf.interval = 0.95, ...)

stat_CosineauMoreySE(mapping = NULL, data = NULL,
  geom = "pointrange", position = "identity", show.legend = NA,
  inherit.aes = TRUE, na.rm = TRUE, conf.interval = 0.95, ...)

.CosiMoreCIs

.CosiMoreSEs
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three options:
If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().
A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data.}

\item{geom}{which geom to use? works only for geoms expecting ymin and ymax.
Default is 'pointrange', 'errorbar' works fine as well.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().}

\item{na.rm}{remove NaNs? Default is TRUE}

\item{conf.interval}{default is 0.95}

\item{...}{further arguments passed to geom}
}
\value{
a ggplot layer with errorbars

Standard layer arguments you probably never need to change (descriptions copied from stat_summary in ggplot2 package)
}
\description{
Cosineau-Morey CIs for within subject factors. For more information, see
Cosineau (2005) and Morey (2008)
\\
\code{stat_CosineauMoreyCI} adds confidence intervals\\
\code{stat_CosineauMoreySE} adds standard errors
}
\details{
see \code{Rmisc::summarySEwithin} for calculations
}
\examples{
# example usage in ggplot
require(ggplot2)
dat <- data.frame(datasets::UCBAdmissions)
ggplot(data = dat, aes(x = Gender, y = Freq, idvar = Dept)) +
       stat_CosineauMoreyCI()

# use errorbar instead
ggplot(data = dat, aes(x = Gender, y = Freq, idvar = Dept)) +
       stat_CosineauMoreyCI(geom='errorbar')

# use SE instead of CI
ggplot(data = dat, aes(x = Gender, y = Freq, idvar = Dept)) +
       stat_CosineauMoreySE(geom='errorbar')

# show difference
ggplot(data = dat, aes(x = Gender, y = Freq, idvar = Dept)) +
       stat_CosineauMoreyCI(alpha=.5, color='red', lwd=3) +
       stat_CosineauMoreySE(alpha=.5, color='blue', lwd=3)

# color by group
ggplot(data = dat, aes(x = Gender, y = Freq, idvar = Dept, color = Gender,
                       group = Gender)) +
       stat_CosineauMoreyCI() +
       ggtitle('95\% Cosineau-Morey within Subject CIs')

}
\seealso{
Rmisc::summarySEwithin, plotSEM, stat_CosineauMoreyCI, plotSEMminmax
}
\author{
Wanja Mössing
}
\keyword{datasets}
