% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.EEG.R
\name{TF_ggplot}
\alias{TF_ggplot}
\title{TF_ggplot}
\usage{
TF_ggplot(
  TF,
  zlim,
  do_interpolate = T,
  contours = F,
  Timecol = "Time",
  Freqcol = "Hz",
  Zcol = "pow",
  Time_unit = "ms",
  Freq_unit = "Hz",
  Z_unit = "dB",
  highlight = NA,
  x_breaks = NA,
  y_breaks = NA,
  precise_time = FALSE
)
}
\arguments{
\item{TF}{a \code{data.table} in long format with columns for at least the
following: Frequencies, Time, power. Must not be "longer" than this; i.e.,
this function cannot handle more than one value per Frequency*Time
combination. So make sure to average over channels before calling this
function.}

\item{zlim}{limits of the colorscale; length 2 numerical vector; e.g., \code{c(-2, 2)}}

\item{do_interpolate}{boolean: Smooth plot via interpolation? (Default: TRUE)}

\item{contours}{boolean: Display contour lines on top? (Default: FALSE)}

\item{Timecol}{name of the column in TF coding time (default: 'Time')}

\item{Freqcol}{name of the column in TF coding frequency (default: 'Hz')}

\item{Zcol}{name of the column in TF containing data (default: 'pow')}

\item{Time_unit}{unit of the time axis, used for label (default: 'ms')}

\item{Freq_unit}{unit of the frequency axis, used for label (default: 'Hz')}

\item{Z_unit}{unit of the z-axis axis, used for label (default: 'dB')}

\item{highlight}{length 4 numeric vector defining the boundaries of the
(optional) rectangle highlighting an area. Default \code{NA} does not show
anything. Otherwise use: \code{c(ymin, ymax, xmin, xmax)}}

\item{x_breaks}{numerical vector with desired breaks on x-axis. Will
approximate and round internally.}

\item{y_breaks}{numerical vector with desired breaks on x-axis. Will
approximate and round internally.}

\item{precise_time}{Due to the smoothing process, there may be rare
imprecisions on the time axis. For instance, the smoothed axis might display
 "1000ms", whereas this datum is actually "1001.2ms". If you set this
 parameter to TRUE, you'll see the true time axis (default: FALSE).}
}
\value{
a \code{ggplot} object
}
\description{
creates a time-frequency plot. Optionally with a highlighted
area and/or contours. Can deal with logarithmic spacing of frequency axis.
Note that this function internally smoothes both axis because TF data from
EEGlab usuallyhave very small numerical imprecisions and are thus not 100%
linear, which irritates ggplot. By default, values are interpolated to
create a smooth visual experience. The plot imitates Matlab's 'jet' colors.
}
\examples{
library(data.table)
TF <- data.table(Time = rep(seq(-500, 3000, 10), 30))
TF[, Hz := 1:30, by = Time][, pow := 0]
TF[Time \%between\% c(2000, 2500) & Hz \%between\% c(7, 14), pow := 2]
TF_ggplot(TF, c(-2,2), highlight = c(7, 14, 2000, 2500))

}
\seealso{
ggplot, geom_raster
}
\author{
Wanja Mössing
}
