% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.permutationtests.R
\name{TimeSeriesClusterBasedPermutationTest}
\alias{TimeSeriesClusterBasedPermutationTest}
\title{Run a paired two sample permutation test for timeseries (e.g., ERPs) with
cluster correction.}
\usage{
TimeSeriesClusterBasedPermutationTest(
  data,
  condCol = "condition",
  dataCol = "mV",
  idCol = "ID",
  timeCol = "Time",
  nperm = 1000,
  altrntv = "two.sided",
  alpha = 0.05,
  cluster.length = 5,
  treatment.group = ""
)
}
\arguments{
\item{data}{a \code{data.table} in long format with columns for at least the
following content}

\item{condCol}{name of the column in data coding the condition labels
(default: 'condition')}

\item{dataCol}{name of the column in data coding the measure values (default
: 'mV')}

\item{idCol}{name of the column in data coding the IDs/Subjects (default:
'ID')}

\item{timeCol}{name of the column in data coding the time (default: 'Time')}

\item{nperm}{number of permutations (default: 1000)}

\item{altrntv}{passed to the t.test functions (default: 'two.sided'); In
case of 'greater' or 'less', you need to specify \code{treatment.group}}

\item{alpha}{significance threshold used for significance detection
(default: 0.05)}

\item{cluster.length}{minimum number of consequtive significant samples to
be considered a significant cluster (default: 5)}

\item{treatment.group}{only necessary if \code{altrntv} is 'greater' or
'less'. Indicates which group is supposed to be greater or less than the
other group. Should be a char matching the label in condCol. (e.g., 'A',
'treatment', 'valid-cue', etc.)}
}
\value{
a \code{data.table} with one row per timepoint. Contains columns for
 cluster-significance, single-sample-significance, t-value of the true
 t-test and permutation p-values. Two additional columns code the beginning
 and end of each cluster.
}
\description{
\code{TimeSeriesClusterBasedPermutationTest} runs a two-sample time-series
permutation test with cluster correction (see Maris and Oostenveldt, 2007).
Specifically, the test runs paired two sample t-tests for each timepoint and
defines cluster as temporally adjacent significant samples, including at least
`cluster.length` samples. It then computes t_sum per cluster (the sum of all
t-statistics of samples involved in that cluster).
Subsequently, `nperm` random partions are created, in which the condition
labels are changed for ~50% of the participants (random without replacement).
For each of these random partitions, t_sum for the previously detected clusters
is computed. Finally, the observed t_sum is compared against the permutation
distribution of t_sums derived from the random partitions, to compute a p-value
per cluster.

Note: for a version without cluster correction, see `TimeSeriesPermutationtest`
}
\details{
alternative = "greater" is the alternative that the treatment.group has a
larger mean than the other group.
}
\examples{
# create fake data
libraries(data.table, ggplot2)
data = data.table()
for (id in 1:10){
  data = rbind(data, data.table(mV = rnorm(4000, 2, 0.5),
                               condition = c(rep('A', 2000), rep('B', 2000)),
                               Time = rep(-500:1499, 2), ID = id))
                               }
# create two "real" clusters around Time 500 & 1000
data[Time \%between\% list(490, 510) & condition == 'A', mV := mV + 10]
data[Time \%between\% list(990, 1010) & condition == 'A', mV := mV - 10]

# plot data
ggplot(data[, .(mV = mean(mV)), by = .(ID, Time, condition)],
       aes(x = Time, y = mV, color = condition)) +
  geom_line()

# run test
foo = TimeSeriesClusterBasedPermutationTest(data, nperm = 100)

}
\seealso{
t.test TimeSeriesPermutationTest
}
\author{
Wanja Mössing
}
