% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.permutationtests.R
\name{TimeSeriesPermutationTest}
\alias{TimeSeriesPermutationTest}
\title{Run a paired two sample permutation test for timeseries (e.g., ERPs)}
\usage{
TimeSeriesPermutationTest(data, condCol = "condition", dataCol = "mV",
  idCol = "ID", timeCol = "Time", nperm = 1000,
  altrntv = "two.sided", alpha = 0.05, cluster.length = 5,
  treatment.group = "")
}
\arguments{
\item{data}{a \code{data.table} in long format with columns for at least the
following content}

\item{condCol}{name of the column in data coding the condition labels
(default: 'condition')}

\item{dataCol}{name of the column in data coding the measure values (default
: 'mV')}

\item{idCol}{name of the column in data coding the IDs/Subjects (default:
'ID')}

\item{timeCol}{name of the column in data coding the time (default: 'Time')}

\item{nperm}{number of permutations (default: 1000)}

\item{altrntv}{passed to the t.test functions (default: 'two.sided'); In
case of 'greater' or 'less', you need to specify \code{treatment.group}}

\item{alpha}{significance threshold used for significance detection
(default: 0.05)}

\item{cluster.length}{minimum number of consequtive significant samples to
be considered a significant cluster (default: 5)}

\item{treatment.group}{only necessary if \code{altrntv} is 'greater' or
'less'. Indicates which group is supposed to be greater or less than the
other group. Should be a char matching the label in condCol. (e.g., 'A',
'treatment', 'valid-cue', etc.)}
}
\value{
a \code{data.table} with one row per timepoint. Contains columns for
 cluster-significance, single-sample-significance, t-value of the true
 t-test and permutation p-values. Two additional columns code the beginning
 and end of each cluster.
}
\description{
\code{TimeSeriesPermutationTest} runs a two-sample time-series
permutation test
Specifically, the test runs paired two sample t-tests for each timepoint and
compares the resulting t-statistics with a permutation distribuion. A minimal
cluster length for consequitive timepoints can be specified (see below.)
}
\details{
alternative = "greater" is the alternative that the treatment.group has a
larger mean than the other group.
}
\examples{
# create fake data
libraries(data.table, ggplot)
data = data.table()
for (id in 1:10){
  data = rbind(data, data.table(mV = rnorm(4000, 2, 0.5),
                               condition = c(rep('A', 2000), rep('B', 2000)),
                               Time = rep(-500:1499, 2), ID = id))
                               }
# create two "real" clusters around Time 500 & 1000
data[Time \%between\% list(490, 510) & condition == 'A', mV := mV + 10]
data[Time \%between\% list(990, 1010) & condition == 'A', mV := mV - 10]

# plot data
ggplot(data[, .(mV = mean(mV)), by = .(ID, Time, condition)],
       aes(x = Time, y = mV, color = condition)) +
  geom_line()

# run test
foo = TimeSeriesPermutationTest(data, nperm = 100)

}
\seealso{
t.test
}
\author{
Wanja Mössing
}
