% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.plaparser.R
\name{tla.parser}
\alias{tla.parser}
\title{Load the output logfile that Nils Lüschow's Android Time Logger App produces,
and parses it to data science friendly long-table format.}
\usage{
tla.parser(
  rawpath,
  fpattern = "^timelogger_data.*\\\\.txt$",
  outpath = NA_character_,
  Hz = 1000
)
}
\arguments{
\item{rawpath}{rawpath to a file or a folder with files (mandatory)}

\item{fpattern}{if path is a folder and regex is not NULL, will compare filenames in path against regex. Default is to match "^timelogger_data.*\\.txt$", which should match all non-renamed files.}

\item{outpath}{path, incl. filename, with output csv (optional)}

\item{Hz}{pseudo-sampling rate of output (default 1000Hz)}
}
\value{
a single data.table in long format with information of all files.
}
\description{
Note: This function was designed to work with V1.1.1 of the app. Since the
      app's output might change at any time, the function may stop working
      with higher versions.
}
\details{
check https://github.com/IXP-Team/exptool_Time-Logger/ to get a release apk.

\code{tla.parser} parses output data from Nils Lüschow's Android Time Logger App
}
\examples{

\dontrun{
DT = tla.parser("raw/", "^f")
}

}
\author{
Wanja Mössing
}
