% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.EEG.R
\name{txt_elec_2_eegUtils}
\alias{txt_elec_2_eegUtils}
\title{txt_elec_2_eegUtils}
\usage{
txt_elec_2_eegUtils(file_name = NULL)
}
\arguments{
\item{file_name}{path+name to the .txt file with spherical coordinates. Default
is our custom Easycap montage (Buschlab WWU Münster).}
}
\value{
a \code{eegUtils} channel location tibble
}
\description{
This is adjusted from \code{eegUtils:::import_chans}, which did
not work for me. Only tested this function with our custom layout file with
the specific constellation used in our lab at the WWU.

Note: depends on Matt Craddock's eegUtils package, which is not available on
CRAN. Use \code{remotes::install_github("craddm/eegUtils")} instead.

If you need your own chanlocs, check the necessary format in the file:
\code{system.file("extdata","easycap_m43v3_biosemi_buschlab_scalponly.txt", package = "wmR")}
}
\examples{
chanlocs <- txt_elec_2_eegUtils()
}
\seealso{
eegUtils eeglabCSV2eegUtils
}
\author{
Wanja Mössing
}
