% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baysCO16ported.R
\name{vonmisespdf}
\alias{vonmisespdf}
\title{vonmisespdf}
\usage{
vonmisespdf(x, mu, K)
}
\arguments{
\item{x}{Theta}

\item{mu}{mu}

\item{K}{Kappa}
}
\description{
Von Mises probability density function (pdf)
 \code{vonmisespdf(x, mu, K)} returns the pdf of the Von Mises
  distribution with mean MU and concentration parameter K,
  evaluated at the values in THETA (given in radians).
}
\author{
Wanja Mössing
}
