% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.R
\docType{package}
\name{wmR-package}
\alias{wmR}
\alias{wmR-package}
\title{wmR: A collection of functions I (Wanja Mössing) frequently use}
\description{
Only use this if you know what you're doing!
}
\section{Ported functions}{

This package contains some ported MATLAB functions.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wamjam/wmR}
  \item Report bugs at \url{https://github.com/wamjam/wmR/issues}
}

}
\author{
\strong{Maintainer}: Wanja Moessing \email{w.a.moessing@gmail.com}

Authors:
\itemize{
  \item Wanja Moessing \email{w.a.moessing@gmail.com}
}

}
\keyword{internal}
