% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.angdiff.R
\name{wmR.angdiff}
\alias{wmR.angdiff}
\title{Take two angles in degree and calculates their difference}
\usage{
wmR.angdiff(deg1, deg2)
}
\arguments{
\item{deg1}{Numeric in range 0:360}

\item{deg2}{Numeric in range 0:360}
}
\value{
a single floating point number representing the difference between the two angles.
  Can take on negative values, as it show the orientation of deg1 IN RELATION to deg2.
  So a result of -1 means, that deg1 is 1° less than deg2.
}
\description{
\code{wmR.angdiff} returns the difference between deg1 and deg2
}
\examples{
#deg1 higher
diff = wmR.angdiff(5, 355)
#deg1 lower
diff = wmR.angdiff(5, 10)

\dontrun{
diff = wmR.angdiff(5000, Inf)
}

}
