% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.R
\docType{package}
\name{wmR}
\alias{wmR}
\alias{wmR-package}
\title{wmR: A collection of functions I (Wanja Mössing) frequently use.}
\description{
Only use this if you know what you're doing!
}
\section{Ported functions}{

This package contains some ported matlab functions...
}

