% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmR.oneliner.R
\name{computername}
\alias{computername}
\title{computername}
\usage{
computername()
}
\value{
Returns character with computername
}
\description{
returns computername; shortcut for \code{Sys.info()['nodename']}
}
\examples{
thispc <- computername()
print(thispc)

if (computername() == 'LinuxServer') {
   rootdir = '/home/wanja/data'
} else if (computername() == 'Laptop') {
   rootdir = 'C:/Users/Wanja/data'
}


}
\author{
Wanja Mössing
}
